requirements.txt
streamlit
gspread
requests
pyzbar
pillow